+++ packages/store/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+# [1.2.0](https://github.com/ngneat/elf/compare/store-1.1.0...store-1.2.0) (2022-02-17)
+
+
+### Features
+
+* **store:** 🔥 add setProps mutation ([452d2aa](https://github.com/ngneat/elf/commit/452d2aa56df7097aff4b387eee97702b0b1c4f24))
+
+
+
+# [1.1.0](https://github.com/ngneat/elf/compare/store-1.0.2...store-1.1.0) (2022-02-16)
+
+
+### Features
+
+* **store:** 🔥 add setProp mutation ([b0bea61](https://github.com/ngneat/elf/commit/b0bea61e67e476559a58ce35018460f8bfac68fb))
+
+
+
 ## [1.0.2](https://github.com/ngneat/elf/compare/store-1.0.1...store-1.0.2) (2021-12-14)